Prevent unsetting non existent values from crashing the app
`git` exits with code 5 when we try to unset a variable that is not set.
Previously this crashed the app and prevent further variables from being
processed. To prevent this, there is now a check to make sure the
variable exist before it is unset. We could instead catch this specific
error, but that would require a more elaborate error handling to
distinguish this from other git errors.